Fixed wrong datatype for motor speed
git-svn-id: https://svn.elec.canterbury.ac.nz/courses/enel429-2011-wackyracers-9@321 3c8fc032-aa9e-0410-b36d-c4f24796240c

+++ src/motor.c
@@ -46,7 +46,7 @@ static void __set_speed(
     // Which (with a little rounding error) results in 0x05A5
     // or:
     //   0x5A << 4 + 0x5A >> 4
-    uint8_t value = (speed << 4) + (speed >> 4);
+    uint16_t value = (speed << 4) + (speed >> 4);
 
     switch (motor + direction) {
         case LEFT+FORWARD:  LEFT_FORWARD_PWM  = value; break;
